pyproject.toml
[project]
name = "satori-ci"
version = "1.5.2"
description = "Satori CI - Automated Software Testing Platform"
authors = [
    {name = "Satori CI CLI", email = "info@satori-ci.com"},
]
dependencies = [
    "requests>=2.27.1",
    "pyyaml>=6.0",
    "flatdict>=4.0.1",
    "rich>=13.0",
    "gitpython>=3.1.31",
    "satori-playbook-validator>=3.2.1",
    "packaging>=23.1",
    "satori-docs",
    "websocket-client>=1.6.1",
    "setuptools>=68.1.2",
]
requires-python = ">=3.9"
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Homepage = "https://www.satori-ci.com"
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
satori-cli = "satoricli:main"
